start figuring out how to load studies into firestore
